% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgen_mod_progression.R
\name{progression_mgen}
\alias{progression_mgen}
\title{Mgen Disease Progression & Recovery}
\usage{
progression_mgen(dat, at)
}
\arguments{
\item{$$}{}
}
\description{
disease progression and recovery
}
